requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0

# LLM providers
openai>=1.57.0
httpx>=0.28.0

# Embeddings & vector search
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0
numpy>=1.26.0

# Configuration
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Linting
ruff>=0.8.0
